python3 str vs. bytes cleanups
sync-client is hitting assorted:
TypeError: a bytes-like object is required, not 'str'

For writing to a file, we can specify the file open in text mode ("w+")
to avoid these.  Similarly for Popen, text=True gives string
compatibility.

There is a Popen for curl to write to icbinn.  That is left as binary
since icbinn_write will take a bytes object.

handle_signal can just encode as utf-8 before writing into the pipe.
The read side uses ord() to pull the int back out and that works fine.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>