sync-cmd: adding ability to create vms
This adds the ability to create vms via the
`unrestricted_create_vm_with_template_and_json` method of xenmgr's unrestricted
interface. This will create a vm based on one of the vm templates stored in
Dom0 while also applying the contents of a JSON file as the vm's configuration.
The JSON file must follow the same conventions as a vm config file stored in
/config/vms.

Signed-off-by: Daniel P. Smith <dpsmith@apertussolutions.com>