sync-cmd: handle names with spaces
VM Names can have spaces, so try to support that.  The xenmgr commands
can just pass the whole strings.  Commands like ~sync-cmd Windows 10
disks` are trickier.

Because of how cmd.py works, we just have a flat string.  Quotes at
invocation time `sync-cmd vm "Windows 10" disks` don't make it through
cmd.py.  We can heuristically parse the names out though.

Through introspection, we can get all valid command names.  Search for a
space prefixed " command" - since the command can't be first, it has to
be prefixed.  If found we can strip the remainder as the command string
and search for the space-containing name.

This means VmCmd has to tuck away the cmd_str for SyncCmd to recover.
This is icky, but it will do the trick.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>