sync-cmd: usb: allow spaces in desc
sync-cmd usb set cannot currently handle spaces in the desc command.
cmd.py is passing in a string, and it doesn't seem possible to get
quotes down into the string.  At least not in any sensible way.

Parse the string in a custom fashion from back to front.  That lets
everything trailing the = to become the "value" and spaces can be
handled.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>